oms-api/requirements.txt
# Production requirements (includes PostgreSQL support)
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0
python-dotenv==1.0.1
python-multipart==0.0.19
slowapi==0.1.9
requests==2.32.3
